Use urfave/cli, add agent command
